requirements.txt
speechrecognition
pyttsx3
pyaudio
llama-cpp-python==0.1.39
customtkinter